Fixed lock order bug in fs.readdir{|One}Helper()
These func's were looping through the response from inode.ReadDir()
that includes "." and "..". Knowing that it was reading a directory,
it had a special case where it compared the dirEntry's InodeNumber
against that of the directory being read. This meant it only matched
".". It would properly skip the step where it would fetch the
InodeLock "." since it already held it... but would curiously go
"ask" what the InodeType for "." was (it must be a Dir of course).

While this unnecessary type fetch for "." was unnecessary, the real
problem was that it would fetch a (fortunately shared) lock on ".."
as part of the "normal" path for non-"." dirEntry's. This is a violation
of the lock ordering requirement and, sure enough, was exposed by
python-swiftclient doing simultaneous GetContainer and DELETE
operations for sufficiently large Containers (i.e. those containing
more objects than the limit, typically 10000, can fetch in a single
query).

Note that the above bug *also* appears in the GetAccount() and even
Readdir() invoked by SMB/NFS/FUSE as those, too, allow for the limiting
of the Readdir() response.