Fix bug that caused Proxyfs to exit when deleting a filesystem.
Fix a bug that caused Proxyfsd to exit, which upsets Samba (smbd),
when a file system is deleting via the controller.

the but is that UnregisterForEvents() is getting called twice for the
same file system, once when it is inactivated and then again when it is
actually deleted.  it should only be called once .

change UnregisterForEvents() so that it does not call logger.Fatalf() when
it is called to unregister a file system that is already unregistered,
instead it just logs an error.

Update associated comments.