Fixed missing {Path|Query}Escape calls in swiftclient
This issue was best revealed when a previously non-ProxyFS Account
that had existing Container and Object data in it was being erased
as part of a `mkproxyfs -F` operation. In the easiest case, creating
an Account, Container, or Object with a "space" in the name would
trigger the Swift server (even ramswift) to not understand the
HTTP Header Line and drop the connection.