Fix two proxyfsd bugs when controller deploys a config.
There are two more bugs where the root cause is proxyfsd getting upset
when the the noauth proxy restarts because existing connections are broken
and new connection attempts fail for a few seconds.

The changes here look larger than they are, mostly because a unit test is
added to try and exercise the bugs and because a portion of the unit test
code is moved to swiftclient/utils.go (the resizeChunkedConnectionPool()
and resizeNonChunkedConnectionPool() routines), which are currently only
called by the swiftclient unit test -- though we may later decide to call
them if a reconfig changes the size of the pools.

Ignoring the changes for the unit tests, which are not executed in a
production environment, the fixes are in acquireChunkedConnection()
and acquireNonChunkedConnection().  there are actually two fixes here,
both of which look like this:

@@ -171,6 +179,16 @@ func acquireChunkedConnection(useReserveForVolumeName string) (connection *conne
        if connectionToBeCreated {
                connection = &connectionStruct{connectionNonce: globals.connectionNonce, reserveForVolumeName: ""}
                err = openConnection("swiftclient.acquireChunkedConnection()", connection)
+               if err != nil {
+                       connection = nil
+                       globals.chunkedConnectionPool.Lock()
+                       globals.chunkedConnectionPool.poolInUse--
+                       globals.chunkedConnectionPool.Unlock()
+
+                       // if openConnection() failed then the cached connections are
+                       // likely broken and using them will create further mischief
+                       drainConnections()
+               }
                stats.IncrementOperations(&stats.SwiftChunkedConnsCreateOps)
        } else {
                stats.IncrementOperations(&stats.SwiftChunkedConnsReuseOps)

This code change fixes two bugs: first, if the openConnection()
fails the code now returns the "connection slot" back to the pool
instead of leaking it, (which eventually causes a hang when all of the
chunked or non-chunked connections are exhaused), and second, when an
openConnection() fails it discards all of the cached connections on
theory that if a new connection can't be established than all cached
connections are actually bad (without this a request might fail because
each of its retries gets a bad cached connection when it tries to retry).

Arguably once we get a connection from the pool and it turns out to be
we should perform a new openConnection() request in-line (which is what
the chunked put code does), but its significant work to retrofit that into
all of the other object, container, and account operations that would need
to be changed so it seems better to simply call drainConnections() here.