Increase the retries for swiftclient for runway and CI.
Increase the "default" values for the number of retries and time allowed
for retries in the the CI and runway environments to ~60 sec.  This should
be more than enough time for the noauth proxy to restart after a reconfig.

For a production environment, the values are set by the proxyfsd config
file(s) which are generated by the Controller or the administrator.
They should be adjusted similarly.

Specifically:
    RetryLimit              5 --> 10
    RetryLimitObject        5 -- > 6
    RetryExpBackoff        1.2 --> 1.4
    RetryExpBackoffObject  2.0 --> no change

For object operations, we now perform 6 retries instead of 5 which
means it will take _at least_ ~63 sec before we give up.  That's
1.0 sec * (2.0^0 + 2.0^1 + 2.01^2 + 2.0^3 + 2.0^4 + 2.0^5) ==> 63 sec

For non-object (container, account) operations, its:
1.0 sec * (1.4^0 + 1.4^1 + 1.4^2 + 1.4^3 + 1.4^4 + 1.4^5 +
           1.4^6 + 1.4^7 + 1.4^8 + 1.4^9) ==> 69.8 sec

This may slow down mkproxyfs for a file system.