Integration of liveness checking package
A JSON response from any alive peer in a cluster to /liveness
is provided that details what has been checked, and by which
"ObservingPeers". Currently, only an RpcPing is issued to the
JSONRPCServer Port on each "ServingPeer". The framework is
ready for enhancements to this checking to also query each
VolumeGroup via their effective VirtualIPAddr/PublicIPAddr
(NFS TCP Port 2049 & (SMB TCP Port 445). As for each Volume,
and to know which (NFS and/or SMB) Port to probe, the liveness
package will eventually need to know the NFSExportName and/or
the SMBShareName for each served Volume.

Note the updating of proxyfsd/mac_cluster_3_peers.conf and the
commensurate new proxyfsd/file_server_mac_3_peers.conf that is
a recipe to run multiple proxyfsd (or liveness) instances on a
single MacOS (leveraging alias's lo0 / localhost interfaces).

TODO: Need to add programatic test validation of the results of
      computeLivenessCheckAssignments()... it has only been
      verified manually (hence the fmt.Println()'s at the end
      of TestComputeLivenessCheckAssignments() and the TODOs
      left as placeholders).