Cannot hold globals.Lock() while up()'ing/down()'ing a Volume
This was causing a deadlock particularly in the down()'ing case.
Here, a final checkpoint is requested. If we then wait for the
checkpoint to complete - while holding the globals.Lock(), it is
possible for a prior checkpoint, having been completed, to attempt
to grab the globals.Lock() [to update some stats] and block. It
will then, therefore, never get back to the point where it reads
this final checkpoint request and, hence, never complete it.