Refactor sundry readdir routines and fix readdir hang.
Write a new fuction, getDirReadLockAndInfo(), which returns with the DLM
lock for the directory held in reader mode and the parent directory's
stat information.  It also performs the permission checks for reading
from a directory.  The stat information for the parent directory is
acquired while avoiding the risk of deadlock and insuring that the parent
directory hasn't changed (the directory could have gotten moved to a
new location any time its not locked).

Collapsed the two readdirHelper routines into a single one that always
returns the stat information for each entry (for readdir plus).
Getting the stat information for an entry is about as expensive as
getting only the type of the inode since both involve locking the inode
and fetching it; the main difference is that getting the stat information
also populates a go map.  But collapsing them means the special handling
for the ".." entry is only done in one place.  readdirHelper() now checks
the error return values that were not being checked and does a bit more
sanity checking of entries read from disk.

Add a rename lock that is held by rename when moving a file or directory
between directories (which means that Rename() has to get the lock on two
arbitrary directories in an undefined order).  GetDirReadLockAndInfo()
holds when getting a lock on the parent directory if try lock fails.

The middleware code has not yet been changed to use the lock.

Between the getDirReadLockAndInfo() function, which gets the directory
lock and performs the permission checks, and the readdirHelper() routine
that does most of the work, the four readdir variants (Readdir(), ReaddirOne(),
ReaddirPlus(), and ReaddirOnePlus()) shrink significantly.