Extend trackedlock checking to jrpcfs and ramswift.
The jrpcfs package was missed in the original conversion of
sync.Mutex --> trackedlock.Mutex.  catch it now.

ramswift already uses trackedlock, but if it was invoked as a
standalone daemon (as jrpcclient tests do), then trackedlocks
cannot be enabled from a config file.  Fix that.  When ramswift
is embedded in another test, it was already possible to enable
lock tracking for that test.