Changed SnapShotPolicy to respond to SIGHUP
Previously, a change to a Volume's SnapShotPolicy specification would
only be picked up on a Volume newly served (including restart of course).
This change makes to interpretation of the SnapShotPolicy occur on SIGHUP
as well.

In addition, this change also halts the taking of SnapShots based on an
existing SnapShotPolicy during SIGHUP handling (that also occurs during
the shutdown sequence).