Bugfix for PFS COALESCE/SLO/S3API multipart-upload for FileCatalyst
pfs_middleware and proxyfs ignored any metdata (headers) provided as
part of a COALESCE operation and did not store them, causing the
newly created object to lack S3 ACLs which meant it could not be
accessed or delted via S3API.

this fix stores any provided headers, other than `X-Static-Large-Object`
and headers that start with `X-Object-Sysmeta-Slo-`.

dropping the headers caused at least one bug--but may have avoided others.
this commit tries to fix the actual bug and the other theoretical bugs.

add code to mung and unmung the several ETag headers that could exist
and insure those ETags are dropped if the file changes via another API.

the bug that this fixes is for FileCatalyst uploads:  if the file being
uploaded already existed and it was created via the `s3api` multipart
upload then FileCatalyst would perform a GET (HEAD?) on the file, and
that would fail with a permission error because the s3Acl that should
have allowed to exceed was dropped on the floor.

Note: files uploaded via other APIs, such as file access or the Swift
object API will also lack an s3Acl and therefore cannot be deleted or
otherwise accessed via `s3api`.  this will likely cause other problems,
but it needs to addressed by a separate fix.

Fix the problem and add some tests to pfs_middleware to verify that
"good" headers are stored and "bad" headers are not stored.

Write bin/aws-multipart-test.py to be able to test the s3api
multipart-upload (SLO) against different accounts and to verify
that ETags and permissions (by implication) are stored correctly.
This test must be run manually and requires some setup.