Bugfixes for pfs middleware PUT involving a directory.
This change includes slight changes to the behavior of the JSON RPC
API used by middleware and also adds an IsDir boolean returned by
RpcGetObject(), which previously failed if invoked on a directory.

Fix MiddlewareGetObject() so it returns a success instead of an error and
does not generate proxyfs log messages if called for an object that is
not a file (previously it tried to return a "read plan" for directories
and sym links and the call to GetReadPlan() would log an error).

Change MiddlewareMkdir() so that it deletes the target object if is a
file and tolerates an existing directory, only updating the metadta,
instead of failing the operation (which caused a lot of problems when
1Space was trying to replicate directories to new nodes).

MiddlewareMkdir() continues to follow the final pathname component if
its a symbolic link, rather than deleting it (which is probably a bug).

Change MiddlewarePutComplete() so that it deletes the target object if
it is an empty directory instead of failing the operation.

MiddlewarePutComplete() continues to follow the final pathname component
if its a symbolic link, rather than deleting it (which is probably a bug).

Fix the bucketstats for MiddlewarePutComplete().  They were
updating the counters for globals.MiddlewareMkdir instead of for
MiddlewarePutComplete. Doh!

Split the code in fs.Rmdir() and fs.Unlink() into
fs.Rmdir()/fs.RmdirActual() and fs.Unlink()/fs.UnlinkActual() so that
the "actual" routines, which do the actual work, can be called by the
MiddlewarePut*() routines if they need to delete something.

Change fs.Unlink() to return IsDirError instead of NotEmptyError when
called on a directory.

Create the resolvePathOption data type and make all of the option
flags type resolvePathOption (i.e. resolvePathFollowDirEntrySymlinks).
Update the type of function arguments

Write unit tests to validate the new behavior of MiddlewarePutComplete()
and MiddlewareGetObject() and also to validate the errors returned
by Rmdir().

Change removeObstacleToObjectPut() to return blunder.NotEmptyError when
a directory is not empty instead of blunder.IsDirError (the error text
is unchanged -- it is still "%s is a non-empty directory").

Change MiddlewareGetObject() to return the inode number as
inode.InodeNumber instead of uint64 for consistency with other API
functions and to improve compile time error checking.

Start work on units tests for MiddlewarePutComplete() and
MiddlewareMkdir().

Wrap some lines longer than 120 characters.