Fixed issues with duplicate bucketstats/transitions registrations
Currently, all transitions registrations happen in init() func's for
each interested client package. These init() func's will be called
precisely once at process launch... before the main() func for the
program is called.

One issue was that a cut&paste error had two transitions registrations
cite the same package name. This was actually innocuous, as it turns out,
but I'd prefer to nevertheless detect this (as was previously done in
bucketstats registrations interestingly). So one change enforces such
uniqueness.

Meanwhile, bucketstats registrations are all made in interested client
package Up() func's... and a corresponding deregistration is made in
those packages' Down() func's. This is actually fine as Up() and Down()
callbacks should be precisely paired. However, there is also an assumption
in the bucketstats registration code that all stats start out zeroed.
This turns out not to be the case. Such an assumption is only guaranteed
at process launch. As such, it was possible... depending on events that
bucketstats captures during a run... for such assumptions to be violated.
This would have the effect similar to a package forgetting to call
the bucketstats deregistration func.

My work-around is to allow re-registrations... and note that bucketstats
are not necessarily starting out at zero.

A better fix would be to eliminate the deregistration in bucketstats and,
instead, move calls to the registration func to each interested client
package's init() func.

Why this is important is because several Go test sequences actually
cause a ProxyFS package stack subset to come Up() and Down() multiple
times breaking the "everything must have started out zeroed" assumption.
This was causing intermittent panic() calls during Make's "go test"
phase... now avoided.

I consider this a TODO to reach a conclusion on how we want this to
ultimately work but, for now, we will avoid these intermittent failures.