Incorporate changes suggested in code review.
Break the functionality of unmung_etags_and_return_best() into
unmung_etags() and return_best_possible_etag() and propagate the changes
through the code.

Along the way, i noticed there were probably some bugs with when and
how we "unmunged" the etags which are fixed.  In addition, post_object()
did not seem to handle the case where the POST added one of the "special"
etags that was not previously present (if it was previously present the
POST would not be able to change it).

put_object_as_directory() now returns an ETag of EMPTY_OBJECT_ETAG which
means the corresponding unit test had to be changed; i think the orginal
behavior was incorrect.

Replace class TestBestPossibleEtag tests with class TestEtagHandling,
which has separate tests for mung_etags(), unmung_etags(),
best_possible_etag().  The tests now run through all three ETag headers.