Further fixes for Etags and random case headers.
A client of pfs_middleware will no longer be able to add
"X-Object-Sysmeta-*" or "X-Static-Large-Object" headers via a POST
operation on an object.

A client of pfs_middleware will no longer be able to slip in a
"X-Object-Sysmeta-ProxyFS-Initial-MD5" header or a
"X-Object-Sysmeta-Proxyfs-Initial-Md5" header to a PUT operation.

Fix a bug where coalesce_object() would fail to return a caller
supplied ETag value.