Correct GetContainer "marker" handling
The mechanism in fs.MiddlewareGetContainer() has to handle the Swift
concepts of "marker", "endmarker", "prefix", and "delimiter". The tense
of particularly "marker" and "prefix" are different however. For "marker",
the returned "objects" are supposed to have names greater than the value
of the marker. For "prefix", the returned "objects" can include an object
matching the value of the prefix. The prior version of this func got these
two contentions slightly mixed up.