Change MiddlewareGetObject() to return `HeadResponse` instead of a bunch of individual values.  This effectively adds IsDir to the values returned by MiddlewareGetObject() so it can be passed through the RPC layer.
Change MiddlewarePutComplete() so that over-writing a "directory object"
with a "file object" succeeds if the directory is empty, and fails if
it is not (previously it would fail even if the directory was empty).
Change the error returned when the PUT fails from from InvalidArgError
to NotEmptyError.  Change the python middleware code to check for
NotEmptyError instead of IsDirError -- IsDirError was never returned, but
we did not have, and still do no have, end-to-end tests which check this.

Add tests that GET on a directory object returns the correct values
and those values match what is returned by HEAD.

Add tests that verify that the "overwrite" via the object API succeeds
provied that either 1) the directory is empty or 2) the directory is
being overwritten with another "directory" (so that only the object
metadata is overwritten).

Add a test that overwriting a "directory object" with a "file object"
succeeds if the directory is empty and fails otherwise.