Make quorum fetching of Checkpoint Header in parallel
Previously, each attempt to fetch the Checkpoint Container's Header
was serialized. With as much as 3 minutes consumed retrying when
receiving a 404 Not Found (as should be expected in an unformatted
Swift Account) and a quorum among 5 attempts, formatting of a new
Swift Account would always take a whopping 15 minutes.

With this change, the above example is accellerated because each of
the 3 minute attempts proceed simultaneously...meaning the quorum
calculation for a new Swift Account would instead only take 3 minutes.