Updated package liveness to workaround hang in SignaledStart().
The issue is (was) SignaledStart() attemps (attempted) to send a
struct{} to (previously unbuffered) stateMachineStopChan...that
would block until do{Candidate|Follower|Leader}() would ultimately
read (as part of a select{} block) causing it so signal WaitGroup
stateMachineDone with Done() and exit. Following the read, the
writer should unblock and proceed to where it does a Wait() on
WaitGroup stateMachineDone and be able to proceed knowing that
stateMachine() has exited. Trouble is, we observed a case where
the stateMachine() goroutine had both issued the Done() call and
exited... but SignaledStart() was still blocked doing the write
of a struct{} to stateMachineStopChan. It is unclear why.

So the fix (workaround) is to always use a 1-element buffered
"stop" chan so that patterns like this will instead only block
on the WaitGroup.Wait() call.