Ensure we checkpoint on explicit flushes only
Previously, we were also calling doInlineCheckpointIfEnabled()
during fs.Wrote calls presumably thinking that this was often
in response to a Swift or S3 PUT request. However, that path
uses fs.MiddlewarePutComplete()... which did not call the
doInlineCheckpointIfEnabled() func btw.

Indeed, if that were the intent, then fully any Swift or S3
request that modifies contents of a BiModal Account should be
calling doInlineCheckpointIfEnabled() as well.

The decision on when/where else to call this func is left for
a more careful review. But the removal of the call from fs.Wrote()
is certainly fine as that API is used by those without RESTful
assumptions (e.g. PFSAgent).