Fixed hang in PFSAgent doing large file writes
The issue occurred whenever a bunch of LogSegment PUTs are still in flight
and they complete out of order (kind of a likely scenario in all but the
most simplified/controlled development/test environments).

The issue turned out to be use of the .Next() API on a Golang container/list.Element that had previously been removed from the List. As you can imagine, some doubly-linked lists remove an item by merely writing the predecessor's Next field to point to the RemovedElement's Next field's value... and writing the successor's Prev field to point to the RemovedElement's Prev field's value. In such a case, the RemovedElement's Next field would still contain a pointer to what was its successor. Golang folks intentionally make this assumption fail (to discourage bad programming behavior) but somehow didn't make execution of .Next() panic or whatever. Go figure...