Minor fix to swiftclient chaos testing.
Change swiftclient/object.go to use atomic.AddUint64() when it
increments the global variable objectFetchChunkedPutContextCnt,
which avoids an unlikely race condition that might occur during
Chaos unit tests.

Move the increment inside an if statement so the atomic operation
is performed only when globals.chaosFetchChunkedPutFailureRate > 0.

this fix has been sitting around my local tree for quite awhile
and I wanted to finish it.