Change confgen to look elsewhere for its templates.
Change the confgen command to look in several places for
"templates/smb_globals.tmpl" and "templates/smb_shares.tmpl":

    /etc/proxyfsd
    /opt/ss/etc/proxyfsd
    .

this is not really the right solution, but overcomes the immediate problem
that confgen is not going to run when invoked by monitord because the
template files that it wants are not going to be in its bin directory.

possibly these belong in a "lib" directory instead of an "etc" directory.