Add bucketstats to checkpoint operations.
Some counters for bytes transferred are incorrect.  That will get
fixed later.

Add time and byte counters to DoCheckpoint() (for synchronous flushes),
checkpointDaemon() (for synchronous and time based flushes), and to
putCheckpoint() which does most of the actual works.

There are additional time based counters in checkpointDaemon() and
putCheckpoint() provided mor granular measurements of the time taken by
various sub-operations.