Track the number of bytes and nodes for BTrees flushed to disk.
Fields are added to bPlusTreeWrapperStruct to track the number
of bytes and number of nodes that are flushed.

PutNode() is modified to add the figures for node being flushed to its
bPlusTreeWrapperStruct and a ClearCounters method is added to clear
the counters.

Code added to putCheckpoing() clears the counters for a tree, flushes
the tree, and then updates the values in the counters to bucketstats.