Changed checkpointDaemon() to batch checkpoint requests.
For a benchmark with 50 threads creating small files using the Fuse
driver (which does a lot of checkpoints) this led to a 5X performance
improvement.

If there are multiple checkpoint requests outstanding, checkpointDaemon()
now performs a single checkpoint and marks all of the requests as done,
whereas previously it would perform a checkpoint for each request.

Delete fetchNextCheckPointDoneWaitGroupWhileLocked() because it was
not called by anyone and seems unnecessary.

Changed the computation of statistics in CheckpointCompleted() to
perform all of the calculations while protected by a lock (move the lock
acquisition up).  Because only one go routine ever calls this function
there was not a bug, but having the correctness of the code depend on
the behavior of the callers is not good.