Performance tweaks for our Fuse filesystem.
The Fuse Flush() routine is invoked whenever close(2) is called for a
file on a proxyfs file system.  There is no need to flush anything in
response to a close().  Change the code from calling Proxyfs Flush(),
which is a heavyweight routine that flushes in progress writes and
performs a checkpoint, to doing nothing.

This gives an additional 2X performance improvement for a workload with
50 threads creating small files, on top of 5X improvement gained by
batching checkpoint requests.

The Proxyfs Readdir() method returns the type of each inode, so there's
no need to call d.mountHandle.GetType() for each inode number in the
list -- the information is already there.