Fixes for IP addresses with netmask.
Change confgen to parse VirutalIpAddr using net.ParseCIDR() to require
a valid CIDR notation IP address and fail the config push otherwise.
It also produces the IP address used for the virtual IP address directory
name in canonical form (in case it wasn't).  (For some reason this
doesn't affect what gets written in ProxyFS config files.)

Changed Proxyfs so it accepts the IP address for VirtualIPAddr in CIDR
form (with a netmask) and without.  Eventually we will figure out where
CIDR is required and where its not, and enforce that, but for now this
should fix the bug that caused ProxyFS to choke on CIDR notation.