Fix bugs in confgen and mkproxyfs.
confen was emitting scripts to make /share directories that had syntax
errors so they wouldn't run (a missing closing `]` in an if statement.
annoyingly, running the script still exited with 0.

mkproxyfs was sending an error message to stdout so etcd-mgmt was not
picking it up.  change it to send the message to stderr.