Bug fixes and changes for confgen.
confgen was using the VirtualIP address of the volume group as the
private IP address used to reach proxyfs.  that is incorrect and probably
doesn't work in general (proxyfs is not going to bind to the virtual IP
addresses, particularly if they are added after Proxyfs starts up).

Added knowledge of `ServingNode` and `PreferredPeer` to confgen.
`ServingNode` overrides `PrimaryPeer` if both are present.  If neither
is present then confgen verifies that `PreferredPeer` is present (can
be empty string) or flags an error in the config.  In the future, the
first time a new volume group is loaded with `monitorctl new-config ...`
ServingNode will not be present (its set by `monitord`) and PrimaryPeer
will not be present (its deprecated), but `PreferredPeer` should be
present in this case.

Add code to confgen to require that `PublicIPAddr` and `PublicInterface`
is present in the config for each peer (node) that is serving a
VolumeGroup as a sanity check.

Fix confgen unit tests to assert that fetchVolumeInfo() did not return
an error (previously unchecked),

Tweak the sample configuration file used by the confgen unit tests
to exercise some of the ServingNode/PrimaryPeer/PreferredPeer code,
and also deleted `PrimaryPeer` from Volume entries.

Add a little logging to proxyfs Format() operation to log the results
of swiftclient.AccountGet().