Change from 'PrimaryPeer' to 'ServingNode'
Change ProxyFS, utilities, and unit tests to use "ServingNode" as
the node that serves a proxyfs file system instead of "PrimaryPeer".

The biggest part of the change is to add a function to the `transitions`
package, `GetServingNode()` that parses a ConfMap and returns the value
for ServingNode (if defined), PrimaryPeer (if defined), or an error.
An empty value results in an empty string (""), which is fine, and
is different than no OptionName at all.

This actually saves a number of lines of code.

Change most of the unit tests to say "SevingNode" instead of "PrimaryPeer"
for the "VolumeGroup" entry and delete the "PrimaryPeer" entry for
Volumes.

Add a unit test for `GetServingNode()` which tries the several
possibilities, including ServingNode and PrimaryPeer are both set,
but to different values (ServingNode wins).