Serialize cascading retransmit requests on server
Serialize retransmit requests on the server by doing:
1. Send UniqueID from client to server when the socket is first open
2. After accepting new socket, the server waits for the UniqueID from
the client.
3. If this is the first time the client has contacted the server, the
server creates a clientInfo structure for the client and proceeds to
wait for RPCs.
4. If this is a client returning on a new socket, the server blocks
until all outstanding RPCs and related goroutines have completed for the
client on the previous connection.
5. Additionally, the server now blocks on accepting new connections
(which could be yet another reconnect for the same client) until the
previous connection has closed down.