Upgrade jQuery due to CVEs reported by GitHub (#494)
WARNING: if you had already built ProxyFS in your development environment, you might need to run `make clean` before trying to build this version of the code (or future versions). You should only need to do it once.
Similarly, you might run into issues if going from this version to an older one. If you need to so so, run `make clean` BEFORE you switch to an older version of the code.