Allow volume groups to not have a virtual IP.
Volume groups that have no volumes shared using SMB or NFS do not need
to have a virtual IP address.  In fact, it is an error if they do (though
I'm not sure it needs to be an error).

Add unit tests.

Fix bug in existing unit tests that failed to notice an error.  Use valid
CIDR IP addresses for unit tests.