Allow a VG that is not be shared to have a VIP.
It is OK for a volume group to be empty (have no volumes).  Therefore we
will allow a VG that does not have any active NFS or SMB shares to have
a Virtual IP address because it may get such a volume (or may have had
such a volume deleted).

It is still rquired that a VG with a volume that is being shared via NFS
or SMB have a VIP.