Add useful startup/log rotate messages to the logger.
When proxyfs or other programs start up, the main function doesn't get a
chance to log "Proxyfsd is starting up" until after package transitions
has run the Up() method for every package because that's how the logger
gets initialized (perhaps we should change that).

Anyway, the logger doesn't know what process/program its part of so it
can't print "Proxyfsd is starting up" because it might be pulled in to
mkproxyfs or some other executable.  But we can at least have the logger
log that it is opening a new log file so its more apparent to someone
reading the log when log files get opened ... and closed.