Fix condition in PFSAgent where DoSetAttr() needs to speculatively reference a fileInode
Unfortunately, the sanity check enforced that fileInode.reference() was only called from DoWrite()
where fileInode.references *WOULD* always be non-zero.

This change switches DoSetAttr() to instead use the referenceFileInode() func to increment the
fileInode.references... possibly from zero. The fact that it will cause a fileInode to be created in
the very small chance it was removed just after DoSetAttr() found it but before the call to
referenceFileInode() is ok... no damage will be done.