Implemented pfsagentd/container/{build|insert}/Dockerfile's
The build/ Dockerfile is provided to construct PFSAgent elements
It is intended that the Container Image resulting from this is
widely available to anyone wanting to buld PFSAgent for a
Linux environment (typically containerized).

The insert/ directory actually includes a bash script and a
sample .conf file that are used to construct a Dockerfile
based off the desired base Docker Image extended to include
the PFSAgent components from the build/ Docker Image as well
as configuration of any desired PFSAgent instances.

It should be noted that the resultant Dockder Image must be
run with `--device /dev/fuse --cap-add SYS_ADMIN` options
since each PFSAgent instance leverages the FUSE subsystem.