Enable, again, a debug HTTP server on localhost:6058.
We need a debug server that starts early in the process and stays running.

Piggy backing on the healthcheck HTTP server doesn't work because it
doesn't start until every other module transitions to Up() and it is
also unavailable during config reloads.  If that hangs, we can't use it
get stack traces.

Changed debug port 6060 --> 6058.