Change serveClient() to use a channel instead of goroutines.
This avoids starting the goroutine repeatedly and should cut down
on time spent growing the stack because, instead, we reuse the
same goroutine.

Additional goroutines are started as needed to insure a request is
never blocked waiting for a goroutine to process it.

Add a debug HTTP server for the unit tests.