Avoid refelction when reading and writing ioHeader.
Add marshalIoHeader() and unmarshalIoHeader() which pack and unpack each
field of the ioHeader struct individually, avoiding the use of reflection
in binary.Read() and binary.Write().

this seems to buy a 1.8% performance improvement, which, when combined
with the increased code complexity and need for the ioHeaderLenOnTheWire
constant makes the utility of this optimization somewhat dubious.