Fix unit tests in "utils".
On the new "github actions" server it appears that some unit tests
can fail because time "stops" for a short period of time (time.Now()
returns the same value on two calls).  Work around the problem.

I also saw a one-time failure in TestTryLockMutex().  Tweak some of the
timing paramaters to try and avoid that in the future even though it
should not have failed in the first place.